new dashboard edit fix

+++ src/main/java/com/facilio/bmsconsole/commands/AddOrUpdateDashboardFieldMappingCommand.java
@@ -26,30 +26,34 @@ public class AddOrUpdateDashboardFieldMappingCommand extends FacilioCommand {
         Long userFilterId = (Long) context.get(FacilioConstants.ContextNames.DASHBOARD_USER_FILTER_ID);
         List<DashboardFieldMappingContext> filterMappings = (List<DashboardFieldMappingContext>) context.get("fieldMappings");
         List<Long> existingIds = new ArrayList<>();
-        if(filterMappings != null && !filterMappings.isEmpty()){
-            for(DashboardFieldMappingContext mappings: filterMappings){
-                if(mappings.getId() <=0){
-                    GenericInsertRecordBuilder insertBuilder = new GenericInsertRecordBuilder()
-                    .table(ModuleFactory.getDashboardFieldMappingModule().getTableName())
-                    .fields(FieldFactory.getDashboardFieldMappingsFields());
-            mappings.setOrgId(AccountUtil.getCurrentOrg().getId());
-            mappings.setDashboardUserFilterId(userFilterId);
-            Map<String, Object> props = FieldUtil.getAsProperties(mappings);
-            insertBuilder.addRecord(props);
-            insertBuilder.save();
-            existingIds.add((Long) props.get("id"));
-        } else {
-            existingIds.add(mappings.getId());
+        if(userFilterId != null && userFilterId > 0){
+            if(filterMappings != null && !filterMappings.isEmpty()){
+                for(DashboardFieldMappingContext mappings: filterMappings){
+                    if(mappings.getId() <=0){
+                        GenericInsertRecordBuilder insertBuilder = new GenericInsertRecordBuilder()
+                                .table(ModuleFactory.getDashboardFieldMappingModule().getTableName())
+                                .fields(FieldFactory.getDashboardFieldMappingsFields());
+                        mappings.setOrgId(AccountUtil.getCurrentOrg().getId());
+                        mappings.setDashboardUserFilterId(userFilterId);
+                        Map<String, Object> props = FieldUtil.getAsProperties(mappings);
+                        insertBuilder.addRecord(props);
+                        insertBuilder.save();
+                        existingIds.add((Long) props.get("id"));
+                    } else {
+                        existingIds.add(mappings.getId());
+                    }
+                }
+            }
+
+            List<Long> ids = DashboardFilterUtil.getFilterMappingIdForFilterId(userFilterId).stream().filter(id -> !existingIds.contains(id)).distinct().collect(Collectors.toList());
+            if(!(ids.isEmpty())){
+                GenericDeleteRecordBuilder builder = new GenericDeleteRecordBuilder()
+                        .table(ModuleFactory.getDashboardFieldMappingModule().getTableName())
+                        .andCondition(CriteriaAPI.getIdCondition(ids,ModuleFactory.getDashboardFieldMappingModule()));
+                builder.delete();
+            }
         }
-    }
-}
-        List<Long> ids = DashboardFilterUtil.getFilterMappingIdForFilterId(userFilterId).stream().filter(id -> !existingIds.contains(id)).distinct().collect(Collectors.toList());
-        if(!(ids.isEmpty())){
-            GenericDeleteRecordBuilder builder = new GenericDeleteRecordBuilder()
-                    .table(ModuleFactory.getDashboardFieldMappingModule().getTableName())
-                    .andCondition(CriteriaAPI.getIdCondition(ids,ModuleFactory.getDashboardFieldMappingModule()));
-            builder.delete();
-        }
+
         return false;
     }
 }
